CourseraAssignment2
This is what I manage to write. I use A and AR as variables. The reason
is I was trying to make it more fun. Since my first initial starts with
A and my last name is AR, I was thinking it would be fun. In addition,
it reminds me the sequence of the program….